feat: fetch data scheduling
