+++ src/main/kotlin/com/relogging/server/domain/plogging/service/PloggingEventService.kt
@@ -41,4 +41,6 @@ interface PloggingEventService {
         item: VolunteeringDetailApiResponseItem,
         url: String,
     ): PloggingEvent
+
+    fun fetchAndSavePloggingEvent()
 }
